+++ changelog.md
@@ -3,6 +3,10 @@ Version 1.2 (XX-XX-2024)
 * I'm back working on this!
 * Decided to skip finishing 1.1 and am just going to 1.2 to have more fun
 * Code being rebassed with functions moved to librarys
+* Update the temp profile adding the ability for the ram/disk temps to also set the fan speed not just the cpu temp
+* Show the PCM Temp
+* Updated the Send command to better allow debug print and returning properly (still needs some work)
+* hopefully cleaned up send command to not go banannas
 
 Version 1.1 (03-26-2024)
 ========================